Set up minimal Go project (#2)
A simple "Hello world" project. 

We use the popular directory structure of putting all binaries in `cmd/` and library code in `pkg/`. To keep the example simple, the binaries are simply hello world in English and Spanish.

Demos internal and external tests, but not benchmarks and examples (even though these are supported).